14. Przeorganizowano projekt w taki sposób aby usunąć klasę Node.java
