+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -38,3 +38,17 @@ public class IntLinkedList {
     }
 
 }
+
+class Node {
+
+    private int value;
+    protected Node prev, next;
+
+    public Node(int i) {
+        value = i;
+    }
+
+    public int getValue() {
+        return value;
+    }
+}
