--- src/main/java/edu/kis/vh/nursery/list/Node.java
@@ -1,15 +1,0 @@
-package edu.kis.vh.nursery.list;
-
-public class Node {
-
-    private int value;
-    protected Node prev, next;
-
-    public Node(int i) {
-        value = i;
-    }
-
-    public int getValue() {
-        return value;
-    }
-}
